refactor: renamed src/utils/regexes.ts to src/utils/index.ts
